requirements.txt
alabaster==0.7.13
asgiref==3.6.0
backend==0.2.4.1
backends==1.4.32
backports.zoneinfo==0.2.1
Beaker==1.12.0
#boss-oidc==1.2.2
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
Django==3.2
django-allauth==0.52.0
django-auto-logout==0.5.1
django-braces==1.15.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-cron==0.6.0
django-crontab==0.7.1
django-filter==22.1
#django-keycloak==0.1.2.dev0
django-keycloak-auth==0.9.5
django-oauth-toolkit==2.2.0
django-oauth2-provider==0.2.6.1
django-oidc==0.1.3
# Editable install with no version control (django-oidc-provider==0.7.0)
#-e /home/erpnext/djangoproject/django-oidc-provider-master
django-openid-auth==0.17
django-rest-framework-keycloak==0.2.0
django-rest-framework-social-oauth2==1.1.0
django-simple-sso==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-oidc-auth==0.8
ecdsa==0.18.0
fdm==0.4.1
filelock==3.9.0
future==0.18.3
idna==2.10
importlib-metadata==6.0.0
itsdangerous==0.24
jwcrypto==1.4.2
keycloak-client==0.15.4
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
mpmath==1.2.1
numpy==1.24.1
oauthlib==3.2.2
oic==0.13.0
opt-einsum==3.3.0
Pillow==9.4.0
platformdirs==3.1.0
plum-dispatch==1.7.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pycryptodomex==3.16.0
pyjwkest==1.4.2
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-jose==3.3.0
python-keycloak==2.11.1
python-keycloak-client==0.2.3
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
scipy==1.10.0
shortuuid==1.0.11
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
sympy==1.11.1
urllib3==1.26.5
virtualenv==20.20.0
webservices==0.7
workos==1.2.0
wrapt==1.14.1
zipp==3.11.0
